- Moved the address alignment test macros for memchr(), memcmp(), memcpy(),   memmove() and memset() into "string_headers.h".

+++ math_logb.c
@@ -59,7 +59,7 @@ __logb(double x)
 {
 	double result;
 
-	result = log(x) / log(FLT_RADIX);
+	result = log(x) / log((double)FLT_RADIX);
 
 	return(result);
 }
